requirements.txt
python-dotenv
requests
xmljson
websockets
pytz
pyairtable
google-genai
deep-translator